Fix AKAHU_ENDPOINT
All of the URL building around the code is using f'{AKAHU_ENDPOINT}/accounts' style format strings, which ends up creating URLs with two slashes like this: `https://api.akahu.io/v1//accounts`

Akahu is now throwing 404 errors for these requests, so we need to either change the config value to not have the slash, or fix all of the URL building. I have chosen to fix the config.